fix: Perbaikan bug pengambilan data dari page sebelumnya yang tidak terupdate

+++ lib/features/beranda/sub_features/detail_katalog/models/detail_model.dart
@@ -3,7 +3,7 @@ class DetailModel {
   final int idMenu;
   final String? keterangan;
   final String? type;
-  final double? harga;
+  final int? harga;
 
   DetailModel({
     required this.idDetail,
